build: Добавил файл config, вынес обработчики в отдельный файл handlers
